refactor(presentation): extract ExecutableUseCase to Application and split runner envelopes
