+++ frontend/src/Presentation/runner/RunnerFailure.ts
@@ -1,0 +1,9 @@
+/**
+ * Failure envelope returned by UseCaseRunner.
+ */
+export type RunnerFailure = {
+    success: false;
+    status: number;
+    code: string;
+    errors?: unknown;
+};
