+++ frontend/src/Presentation/runner/RunnerSuccess.ts
@@ -1,0 +1,11 @@
+
+/**
+ * Success envelope returned by UseCaseRunner.
+ *
+ * @template TData Payload type of successful response.
+ */
+export type RunnerSuccess<TData> = {
+    success: true;
+    status: 200;
+    data: TData;
+};
